Moved category to product module, renamed rpoduct module to catalog
