+++ core/modules/catalog/components/ProductAttribute.ts
