+++ core/modules/catalog/components/ProductGallery.ts
