+++ core/modules/catalog/components/ProductTile.ts
