+++ core/modules/catalog/helpers/index.ts
