+++ core/modules/catalog/helpers/optionLabel.ts
