+++ core/modules/catalog/queries/searchPanel.js
