+++ core/modules/catalog/store/category/getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex'
 import RootState from '@vue-storefront/store/types/RootState'
-import CategoryState from '../types/CategoryState'
+import CategoryState from '../../types/CategoryState'
 
 const getters: GetterTree<CategoryState, RootState> = {
   current: (state) => state.current,
