+++ core/modules/catalog/store/category/mutations.ts
@@ -3,7 +3,7 @@ import { MutationTree } from 'vuex'
 import * as types from './mutation-types'
 import { slugify, breadCrumbRoutes } from '@vue-storefront/store/helpers'
 import { entityKeyName } from '@vue-storefront/store/lib/entities'
-import CategoryState from '../types/CategoryState'
+import CategoryState from '../../types/CategoryState'
 import rootStore from '@vue-storefront/store'
 
 const mutations: MutationTree<CategoryState> = {
