+++ core/modules/catalog/store/product/index.ts
