+++ core/modules/catalog/store/stock/index.ts
