+++ core/modules/catalog/types/AttributeState.ts
