+++ core/modules/catalog/types/CategoryState.ts
