+++ core/modules/catalog/types/ProductState.ts
