+++ core/modules/catalog/types/StockState.ts
