+++ src/themes/default/components/core/ProductAttribute.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import { ProductAttribute } from '@vue-storefront/core/modules/product/components/ProductAttribute.ts'
+import { ProductAttribute } from '@vue-storefront/core/modules/catalog/components/ProductAttribute.ts'
 
 export default {
   mixins: [ProductAttribute]
