+++ src/themes/default/components/core/blocks/Product/Related.vue
@@ -19,7 +19,7 @@
 <script>
 import ProductListing from 'theme/components/core/ProductListing'
 
-import { prepareRelatedQuery } from '@vue-storefront/core/modules/product/queries/related'
+import { prepareRelatedQuery } from '@vue-storefront/core/modules/catalog/queries/related'
 import i18n from '@vue-storefront/i18n'
 import store from '@vue-storefront/store'
